Tech Week Challenge 2
Ramesh loves mathematics a lot. He always spends his time by playing with arrays.
Suddenly,He stuck with the hard problem. The problem is listed below.
Help, Ramesh to solve this magical problem of mathematics.
Given an array of integers, rearrange the array in such a way that the 
first element is first maximum and second element is first minimum and so on. 
You will be given a function which will take integer array as an argument.
Return the each element of the array after alternate sorting from the function in the form of integer array.

Sample Input :
7
1
2
3
4
5
6
7

Sample Output :
7
1
6
2
5
3
4